#3 Generically flattens a given List-Map structure
An important aspect of the create-Repository script is the modification
of the imported JSON configuration. Nested objects are put into arrays
where only one object resides.

In order to gain a future-safe implementation the current script is
refactored for recursive modification as it was before. What actually
changed is the way the map creation is done: Previously it was hardcoded
with identifiers which was unsatisfactory in terms of future changes.

Instead of hardcoded identifiers, the type is now considered.

Since script debugging on a nexus host is a certain kind of complication,
a special execption is thrown to indicate that the repository may
already exist.

The create-repository scripts was also refactored according its call
hierarchy. Previously there was no pattern recognizable how the methods
call each other.