#3 Adds generic data structure support for Nexus 3 in go, cleans groovy
Modifying the groovy scripts for dealing with creating repositories was
a bad decision as the debugging support is overall bad. Amending the
data structure is necessary for a couple of reasons:

1. Reduce redundancy
  - the repository name should not exist 3x in a single repository
  definition
  - `id`, `repository`, and the repository ID itself
2. Meta data must be added to repository configuration
  - missing meta might prevent further administration of repositories
  created this way. This includes also the deletion.
  - also from the reduce-the-redundancy department:
    - the recipe name holds the type and format of the repository that
    can never change
    - `type` and `format` must be set explicitly but can be derived from
    directly from the recipe name. The user should not be in charge of
    this in order to remove misconfiguration
3. Add support for arbitrarily nested repository configurations
  - only supported with nexus 3

With close to no debugging support inside of groovy-nexus scripts inside
the nexus script, it is better to keep the groovy files as small and
featureless as possible. This is another reason to solve the whole
configurations within the Go files.

- Adds custom example `.hcl` and complimentary remove `.hcl` and their
respective json output. This example file contains also repositories
that have a deeper nesting level than 2. This is an important change
from nexus 2 to nexus 3.

Please note, that nexus 2 will not be supported for deeper nesting
levels.